Make the Ex command's flags booleans
`ExArg` carried `skip`, `forceit`, `append`, `usefilter`, `read_edit` and
`mkdir_p` as `c_int`, which is what C had to spell them as. Each is
written only ever as 0 or 1, so each is a `bool`, and `!= 0`/`as c_int`
goes with it.

`force_bin` stays a `c_int` and is **not** in that list, although the
survey put it there: it is three-valued (`0`, `FORCE_BIN`, `FORCE_NOBIN`),
because `++bin` and `++nobin` have to be distinguishable from "neither was
given". The unit test added with the oracle pins that.

`check_readonly` took the flag as `*mut c_int` so the confirmation dialog
could set it; it takes `&mut bool` and is no longer an `unsafe fn`.

+++ crates/nvim/src/buffer/info.rs
@@ -140,7 +140,7 @@ pub fn buflist_list(excmd: &mut ExArg) {
         if got_int.get() {
             break;
         }
-        if skip(buf, arg, forceit) {
+        if skip(buf, arg, c_int::from(forceit)) {
             continue;
         }
         // The name to show, then the message filter's verdict on it:
